Add a trick iterator
Create a simple solution for a trick iterator. It’s only responsible for validating and choosing the winner among plays.

It uses String as data structure to be able to quickly evolve the idea. However it can easily be changed.

By using TDD, handle different possible plays for a trick, including invalid ones.

The validation is done using Swift error handling, because it makes easier to understand why a play is invalid since it has multiple causes.

Remove files created by XCode when creating the project and the personal folder `xcuserdata`(also ignore this one).