Turn Play to [Card]
Introduce Rank enum to take advantage of Swift’s type safety

Extending Card to conform to ExpressibleByStringLiteral is a trade-off since it looses the type safety again. However, since it’s on tests module, this downside is only for tests and we’re using fatalError to easily find out when a invalid String is used to create a Card.