Refactor trick rules validation
Creating a separated class with a function for each rule, makes it more readable. This version was chosen but I also tested another option.

The second option considered was creating a class for each rule and adding an array of rules in TrickIterator. That way it would be open for extension. However, we wouldn’t have control on new rules and those could *conflict with existing ones*.

Although in future versions we probably wanna different set of rules, based on user’s preference, we’ll keep this version readable, but simple.