Add PlayersSorter
Go simple with dictionary as data structure to evolve the idea quickly. However, typealias and helper functions like makePlayersRoles were added to make the transaction to a new type easy.

A real simple sort approach for this version. It seemed like the easiest and more readable way for now.

Let’s start with 5 players only so we don’t need to bother with all variants. In the future, changing some validation rules will make it possible to have different number of players.

Still, everything done with TDD.